SLING-9636 - Refine feature model files
Move repoinit to the base feature. Using the base feature without repoinit
does not work, and apparently there is a race condition (or ordering problem)
where the jcr installer starts and is unable to find a service user mapping.

This was exposed when renaming the files and potentially altering startup/
processing order.