fix: don't use the default local Maven repository
on the ASF Jenkins instance we configure a per-workspace repository which is used to download
remote artifacts and $HOME/.m2/repository is not touched. This means that old SNAPSHOT artifacts
held in that directory are never refreshed.

For now, work around by passing a build-local repository url to minimise downloading and avoid the
global $HOME/.m2/repository directory.